pyproject.toml
[tool.poetry]
name = "bfa"
version = "0.1.0"
description = ""
authors = ["Sobrino Julian <sobrinojulian@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
python-dotenv = "^0.19.2"
htmlmin = "^0.1.12"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.11b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
